+++ herostratus/src/rules/impls/mod.rs
@@ -13,6 +13,7 @@ mod h013_fortune_teller;
 mod h014_added_first_ci;
 mod h015_octopus;
 mod h016_cthulhu;
+mod h017_ex_nihilo;
 mod tentacle_merge_config;
 
 pub use h002_shortest_subject::H002Config;