+++ CHANGELOG.md
@@ -19,6 +19,7 @@ focus on the user impact** rather than the actual changes made.
   `rules.tentacle_merge.cthulhu_threshold`.
 * New `H16-cthulhu` rule: granted to the author of a merge commit with at least
   `rules.tentacle_merge.cthulhu_threshold` parents.
+* New `H17-ex-nihilo` rule: granted to the author of an empty root commit
 * Rules are now versioned, giving me the ability to trigger rerunning rules on the full history of a
   project when the rule (or its cache) changes.